requirements.txt
# Core pipeline dependencies
pandas==2.2.2
psycopg2-binary==2.9.9
numpy==1.26.4
sqlalchemy==2.0.30

# Great Expectations (OSS, minimal install)
great-expectations==0.18.15

# Testing
pytest==8.2.0
pytest-cov==5.0.0

# Airflow extras (provider packages needed for local installs / tests)
# Note: apache-airflow itself is installed via the base Docker image.
# List here only the packages needed when running tests outside Docker.
apache-airflow==2.9.2
